refactor: move url parsing logic to mapper
