+++ libs/file/src/file.service.ts
@@ -11,8 +11,6 @@ import {
   InternalServerErrorException,
   Logger,
 } from '@nestjs/common';
-import { File, FileType } from '@generated/prisma/client';
-import { DocumentDto } from 'apps/api/src/notice/dto/res/generalNotice.dto';
 
 @Injectable()
 @Loggable()
@@ -91,25 +89,7 @@ export class FileService {
     await Promise.all(keys.map((key) => this.deleteFile(key)));
   }
 
-  getFilesUrl(files: File[]): {
-    imageUrls: string[];
-    documents: DocumentDto[];
-  } {
-    const imageUrls: string[] = [];
-    const documents: DocumentDto[] = [];
-    for (const file of files) {
-      if (file.type === FileType.IMAGE) {
-        imageUrls.push(`${this.s3Url}${file.url}`);
-      }
-      if (file.type === FileType.DOCUMENT) {
-        documents.push(
-          new DocumentDto({
-            url: file.url,
-            name: file.name,
-          }),
-        );
-      }
-    }
-    return { imageUrls, documents };
+  getFilesUrl(url: string): string {
+    return `${this.s3Url}${url}`;
   }
 }
